% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jp_countries.R
\docType{data}
\name{jp_countries}
\alias{jp_countries}
\title{All countries.}
\format{A data frame with 238 rows and 23 variables:
\describe{
  ...
}}
\source{
\url{https://joshuaproject.net/resources/datasets}
}
\usage{
jp_countries
}
\description{
A dataset containing a listing of countries with demographic aggregates. Attributes can be
found in the data frame \code{jp_field_names} (See rows where
\code{jp_field_names$TableName == "jpcountries"}).
}
\keyword{datasets}
