% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jp_field_names.R
\docType{data}
\name{jp_field_names}
\alias{jp_field_names}
\title{General field descriptions.}
\format{A data frame with 231 rows and 4 variables:
\describe{
  ...
}}
\source{
\url{https://joshuaproject.net/resources/datasets}
}
\usage{
jp_field_names
}
\description{
A dataset containing a listing of field names for the various tabular datasets distributed
by the Joshua Project.
}
\details{
The \code{joshuaproject} package includes the following tables:
\describe{
  \item{\code{jpcountries}}{See the \code{jp_countries} data frame.}
  \item{\code{jplangpeopctry}}{See the \code{jp_langpeopctry} data frame.}
  \item{\code{jppeoples}}{See the \code{jp_peoples} data frame.}
  \item{\code{jpupgotd}}{See the \code{jp_upgotd} data frame.}
  \item{\code{jplanguages}}{See the \code{jp_languages} data frame.}
}
}
\keyword{datasets}
