% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jp_langpeopcrty.R
\docType{data}
\name{jp_langpeopctry}
\alias{jp_langpeopctry}
\title{People group-country-language table.}
\format{A data frame with 41605 rows and 7 variables:
\describe{
  ...
}}
\source{
\url{https://joshuaproject.net/resources/datasets}
}
\usage{
jp_langpeopctry
}
\description{
A dataset containing a listing of people groups by language and country. Attributes can be
found in the data frame \code{jp_field_names} (See rows where
\code{jp_field_names$TableName == "jplangpeopctry"}).
}
\keyword{datasets}
