% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jp_languages.R
\docType{data}
\name{jp_languages}
\alias{jp_languages}
\title{All language data.}
\format{A data frame with 7107 rows and 20 variables:
\describe{
  ...
}}
\source{
\url{https://joshuaproject.net/resources/datasets}
}
\usage{
jp_languages
}
\description{
A dataset containing a listing of linguistic groups.  Attributes can be found in the data frame
\code{jp_field_names} (See rows where \code{jp_field_names$TableName == "jplanguages"}).
}
\keyword{datasets}
