% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jp_peoples.R
\docType{data}
\name{jp_peoples}
\alias{jp_peoples}
\title{All people groups by country.}
\format{A data frame with 16818 rows and 33 variables:
\describe{
  ...
}}
\source{
\url{https://joshuaproject.net/resources/datasets}
}
\usage{
jp_peoples
}
\description{
A dataset containing a listing of people groups organized by the country in which they reside.
Attributes can be found in the data frame \code{jp_field_names} (See rows where
\code{jp_field_names$TableName == "jppeoples"}).
}
\keyword{datasets}
