% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jp_upgotd.R
\docType{data}
\name{jp_upgotd}
\alias{jp_upgotd}
\title{Only unreached people groups by country.}
\format{A data frame with 6956 rows and 33 variables:
\describe{
  ...
}}
\source{
\url{https://joshuaproject.net/resources/datasets}
}
\usage{
jp_upgotd
}
\description{
A dataset containing a listing of people groups deemed "unreached" by the Joshua Project,
organized by the country in which they reside. Attributes can be found in the data frame
\code{jp_field_names} (See rows where \code{jp_field_names$TableName == "jpupgotd"}).
}
\keyword{datasets}
